cleanup timestamp management
