+++ src/hal/macros.h
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <string.h>
 
-//#define DEBUG
+#define DEBUG
 
 #define HAL_DANGER(mod, x, ...) \
     do { \
